add basic reconnect structure for our endpoint disconnection failure
