+++ app/src/main/java/com/example/pabloandtyler/comp512app/TextFight.java
@@ -53,6 +53,7 @@ public class TextFight extends AppCompatActivity
     private HashMap<String, String> peersMap = null; //maps endpointId to friendly names
     private HashMap<String, String> peersColorMap = null; //maps endpointId to an assigned color
     private String[] colors = null;
+    private String myFriendlyName;
 
     // Our handle to Nearby Connections
     private ConnectionsClient connectionsClient;
@@ -61,6 +62,7 @@ public class TextFight extends AppCompatActivity
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_text_fight);
+
         connectionsClient = Nearby.getConnectionsClient(this);
 
 
@@ -213,6 +215,7 @@ public class TextFight extends AppCompatActivity
 
     /** Starts looking for other players using Nearby Connections. */
     private void startDiscovery() {
+
         Log.i(TAG, "startDiscovery: finding peers");
         // Note: Discovery may fail. To keep this demo simple, we don't handle failures.
         connectionsClient.startDiscovery(
@@ -221,6 +224,7 @@ public class TextFight extends AppCompatActivity
                         new OnSuccessListener<Void>() {
                             @Override
                             public void onSuccess(Void aVoid) {
+                                Log.i(TAG, "onSuccess: " + getPackageName());
                                 Log.i(TAG, "We're looking for someone to play against");
                             }
                         }
@@ -242,9 +246,10 @@ public class TextFight extends AppCompatActivity
     // Callbacks for finding other devices
     private final EndpointDiscoveryCallback endpointDiscoveryCallback =
             new EndpointDiscoveryCallback() {
+
+                //find peers and update GUI here
                 @Override
                 public void onEndpointFound(String endpointId, DiscoveredEndpointInfo info) {
-                    //find peers and update GUI here
 
                     Nearby.getConnectionsClient(getApplicationContext()).requestConnection(
                             CodenameGenerator.generate(),
@@ -280,7 +285,6 @@ public class TextFight extends AppCompatActivity
                 }
 
                 @Override
-                //TODO: add logic depending on the situation of losing a peer
                 public void onEndpointLost(String endpointId) {
                     Log.i(TAG, "endpoint lost");
                 }
@@ -295,7 +299,7 @@ public class TextFight extends AppCompatActivity
                 Toast.makeText(TextFight.this, "accepting peer?", Toast.LENGTH_SHORT).show();
 
                 if(mode.equals(MainActivity.MODE_HOST)){
-                    //As a host, auto accept an incoming connection
+                    // In host mode, auto accept an incoming connection
 
                     Nearby.getConnectionsClient(TextFight.this).
                             acceptConnection(endpointId, payloadCallback);
@@ -307,15 +311,16 @@ public class TextFight extends AppCompatActivity
                 }
 
                 else if(mode.equals(MainActivity.MODE_PEER)){
-                    //TODO: add new peers to list but do not auto-accepet
+
+                    // add new peers to list of found peers, but do not auto-accept
                     peerListItemsFragment.insertPeer(
                             connectionInfo.getAuthenticationToken(),
                             endpointId,
                             connectionInfo.getEndpointName());
 
-                    Log.i(TAG, "onConnectionInitiatd, MODE = PEER");
-                }
+                    Log.i(TAG, "onConnectionInitiated, MODE = PEER");
 
+                }
             }
 
             @Override
@@ -328,15 +333,17 @@ public class TextFight extends AppCompatActivity
 
                         Toast.makeText(TextFight.this, "accepted peer!", Toast.LENGTH_SHORT).show();
                         break;
+
                     case ConnectionsStatusCodes.STATUS_CONNECTION_REJECTED:
                         // The connection was rejected by one or both sides.
                         Toast.makeText(TextFight.this, "peer rejected", Toast.LENGTH_SHORT).show();
                         Log.i(TAG, "onConnectionResult: connection failed");
                         break;
+
                     case ConnectionsStatusCodes.STATUS_ERROR:
                         //the connection broke before it was able to be accepted
-                        //TODO: if we cannot connect to a peer, have logic here to indicate such
                         Log.e(TAG, "status error for onConnectionResult");
+                        Toast.makeText(getApplicationContext(), "Error happened", Toast.LENGTH_SHORT).show();
                         break;
                 }
 
@@ -344,9 +351,11 @@ public class TextFight extends AppCompatActivity
 
             @Override
             public void onDisconnected(String endpointId) {
-                //TODO: depending on the type of disconnection and topology, have additional logic to conform to situation
+
                 Log.i(TAG, "onDisconnected: disconnected from the opponent");
                 Toast.makeText(TextFight.this, "disconnected", Toast.LENGTH_SHORT).show();
+
+                attemptReconnection(endpointId);
             }
         };
 
@@ -382,7 +391,7 @@ public class TextFight extends AppCompatActivity
     @Override
     public void onPeerClicked(PeerDataItem item) {
         //TODO: initialize connection here with new peer, and display alert dialog.
-        //Toast.makeText(this, "Peer clicked", Toast.LENGTH_LONG).show();
+
 
         JoinPeerAlert alert = JoinPeerAlert.newInstance(item);
         alert.setmListener(this);
@@ -398,6 +407,9 @@ public class TextFight extends AppCompatActivity
         Nearby.getConnectionsClient(this)
                 .acceptConnection(item.getEndpointId(),payloadCallback);
 
+        // allow others to connect to us automatically now that we're in the network
+        mode = MainActivity.MODE_HOST;
+
         //update a new peer to our peers map
         peersMap.put(item.getEndpointId(), item.getFriendlyName());
 
@@ -408,8 +420,9 @@ public class TextFight extends AppCompatActivity
         fragmentManager.beginTransaction()
                 .replace(R.id.multi_fragments,
                         textMainArenaFragment)
-                .addToBackStack("PeerListItemFragment")
+                .addToBackStack("PeersListFragment")
                 .commit();
+
 
     }
 
@@ -422,6 +435,7 @@ public class TextFight extends AppCompatActivity
     @Override
     public void onTextMainFragmentInteraction(String message) {
         Log.i(TAG, "received callback: message: " + message);
+
         for(String endpointId: peersMap.keySet()){
             Log.i(TAG, "sending " + message + " to " + peersMap.get(endpointId));
             sendPayload(endpointId, message);
@@ -466,7 +480,7 @@ public class TextFight extends AppCompatActivity
                 .addToBackStack("TextFightArenaFragment")
                 .commit();
     }
-
+/*
     private class PeerBattleInfo{
         private String peerColor;
         private String peerFriendlyName;
@@ -505,7 +519,40 @@ public class TextFight extends AppCompatActivity
             this.endpointId = endpointId;
         }
 
+    }*/
 
+    private void attemptReconnection(String endpointId){
+        Toast.makeText(getApplicationContext(), "reconnect!", Toast.LENGTH_SHORT).show();
+
+        //reattempt connection only once
+        Nearby.getConnectionsClient(getApplicationContext()).requestConnection(
+                CodenameGenerator.generate(),
+                endpointId,
+                connectionLifecycleCallback)
+                .addOnSuccessListener(
+                        new OnSuccessListener<Void>() {
+                            @Override
+                            public void onSuccess(Void aVoid) {
+                                //We successfully requested a connection. Now both
+                                // sides must accept before the connection is established
+                                Toast.makeText(TextFight.this,
+                                        "onSuccess: requested connection",
+                                        Toast.LENGTH_SHORT).show();
+                                Log.i(TAG, "requested connection: both must accept");
+                            }
+                        }
+                )
+                .addOnFailureListener(
+                        new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+
+                                Log.e(TAG, "onFailure: " + e.getClass() + " " + e.getMessage());
+                                Log.e(TAG, "will stop trying action of reattempts");
+
+                            }
+                        }
+                );
 
     }
 }
