+++ app/src/main/res/layout/activity_text_fight.xml
